refactor(record-field): one export per guard, schemas to validation-schemas
Composite/value guard files defined their zod schema inline — some
exported it (two exports), others kept it private. Move every schema into
validation-schemas/ as a single-export file (matching the existing
currency/select schemas) so each guard file only imports its schema and
exports the guard:

- arraySchema               -> arrayFieldValueSchema
- emailsSchema              -> emailsFieldValueSchema
- filesSchema               -> filesFieldValueSchema
- linksSchema               -> linksFieldValueSchema
- phonesSchema              -> phonesFieldValueSchema
- jsonWithoutLiteralsSchema -> rawJsonFieldValueSchema
- richTextSchema            -> richTextFieldValueSchema
- currencySchema            -> currencyFieldValueSchema
- fullnameSchema            -> fullNameFieldValueSchema

After this no file in types/guards/ imports zod or has more than one
export. Field inputs and the phones settings form import the schema from
its new location. No behavior change.
