+++ packages/twenty-front/src/modules/object-record/record-field/ui/meta-types/input/components/FilesFieldInput.tsx
@@ -9,7 +9,7 @@ import { MULTI_ITEM_FIELD_INPUT_DROPDOWN_ID_PREFIX } from '@/object-record/recor
 import { uploadMultipleFiles } from '@/object-record/record-field/ui/meta-types/utils/uploadMultipleFiles';
 import { recordFieldInputIsFieldInErrorComponentState } from '@/object-record/record-field/ui/states/recordFieldInputIsFieldInErrorComponentState';
 import { type FieldFilesValue } from '@/object-record/record-field/ui/types/FieldMetadata';
-import { filesSchema } from '@/object-record/record-field/ui/types/guards/isFieldFilesValue';
+import { filesFieldValueSchema } from '@/object-record/record-field/ui/validation-schemas/filesFieldValueSchema';
 import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
 import { filePreviewState } from '@/ui/field/display/states/filePreviewState';
 import { useAtomStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomStateValue';
@@ -39,7 +39,7 @@ export const FilesFieldInput = () => {
 
   const parseFilesArrayToFilesValue = useCallback(
     (filesArray: FieldFilesValue[]) => {
-      const parseResponse = filesSchema.safeParse(filesArray);
+      const parseResponse = filesFieldValueSchema.safeParse(filesArray);
       if (parseResponse.success) {
         return parseResponse.data;
       }
