+++ packages/twenty-front/src/modules/object-record/record-field/ui/types/guards/isFieldArrayValue.ts
@@ -1,8 +1,7 @@
 import { type FieldArrayValue } from '@/object-record/record-field/ui/types/FieldMetadata';
-import { z } from 'zod';
-
-export const arraySchema = z.union([z.null(), z.array(z.string())]);
+import { arrayFieldValueSchema } from '@/object-record/record-field/ui/validation-schemas/arrayFieldValueSchema';
 
 export const isFieldArrayValue = (
   fieldValue: unknown,
-): fieldValue is FieldArrayValue => arraySchema.safeParse(fieldValue).success;
+): fieldValue is FieldArrayValue =>
+  arrayFieldValueSchema.safeParse(fieldValue).success;
