+++ packages/twenty-front/src/modules/object-record/record-field/ui/types/guards/isFieldFilesValue.ts
@@ -1,26 +1,7 @@
 import { type FieldFilesValue } from '@/object-record/record-field/ui/types/FieldMetadata';
-import { z } from 'zod';
+import { filesFieldValueSchema } from '@/object-record/record-field/ui/validation-schemas/filesFieldValueSchema';
 
-const fileSchema = z.object({
-  fileId: z.string(),
-  label: z.string(),
-  extension: z.string().optional(),
-  url: z.string().optional(),
-  fileCategory: z
-    .enum([
-      'ARCHIVE',
-      'AUDIO',
-      'IMAGE',
-      'PRESENTATION',
-      'SPREADSHEET',
-      'TEXT_DOCUMENT',
-      'VIDEO',
-      'OTHER',
-    ] as const)
-    .optional(),
-});
-
-export const filesSchema = z.array(fileSchema);
 export const isFieldFilesValue = (
   fieldValue: unknown,
-): fieldValue is FieldFilesValue[] => filesSchema.safeParse(fieldValue).success;
+): fieldValue is FieldFilesValue[] =>
+  filesFieldValueSchema.safeParse(fieldValue).success;
