+++ packages/twenty-front/src/modules/object-record/record-field/ui/types/guards/isFieldFullNameValue.ts
@@ -1,13 +1,7 @@
-import { z } from 'zod';
-
 import { type FieldFullNameValue } from '@/object-record/record-field/ui/types/FieldMetadata';
-
-const fullnameSchema = z.object({
-  firstName: z.string(),
-  lastName: z.string(),
-});
+import { fullNameFieldValueSchema } from '@/object-record/record-field/ui/validation-schemas/fullNameFieldValueSchema';
 
 export const isFieldFullNameValue = (
   fieldValue: unknown,
 ): fieldValue is FieldFullNameValue =>
-  fullnameSchema.safeParse(fieldValue).success;
+  fullNameFieldValueSchema.safeParse(fieldValue).success;
