+++ packages/twenty-front/src/modules/object-record/record-field/ui/types/guards/isFieldLinksValue.ts
@@ -1,19 +1,7 @@
-import { z } from 'zod';
 import { type FieldLinksValue } from '@/object-record/record-field/ui/types/FieldMetadata';
-
-export const linksSchema = z.object({
-  primaryLinkLabel: z.string().nullable(),
-  primaryLinkUrl: z.string().nullable(),
-  secondaryLinks: z
-    .array(
-      z.object({
-        label: z.string().nullable(),
-        url: z.string().nullable(),
-      }),
-    )
-    .nullable(),
-}) satisfies z.ZodType<FieldLinksValue>;
+import { linksFieldValueSchema } from '@/object-record/record-field/ui/validation-schemas/linksFieldValueSchema';
 
 export const isFieldLinksValue = (
   fieldValue: unknown,
-): fieldValue is FieldLinksValue => linksSchema.safeParse(fieldValue).success;
+): fieldValue is FieldLinksValue =>
+  linksFieldValueSchema.safeParse(fieldValue).success;
