+++ packages/twenty-front/src/modules/object-record/record-field/ui/types/guards/isFieldRichTextValue.ts
@@ -1,12 +1,7 @@
 import { type FieldRichTextValue } from '@/object-record/record-field/ui/types/FieldMetadata';
-import { z } from 'zod';
-
-export const richTextSchema: z.ZodType<FieldRichTextValue> = z.object({
-  blocknote: z.string().nullable(),
-  markdown: z.string().nullable(),
-});
+import { richTextFieldValueSchema } from '@/object-record/record-field/ui/validation-schemas/richTextFieldValueSchema';
 
 export const isFieldRichTextValue = (
   fieldValue: unknown,
 ): fieldValue is FieldRichTextValue =>
-  richTextSchema.safeParse(fieldValue).success;
+  richTextFieldValueSchema.safeParse(fieldValue).success;
