+++ packages/twenty-front/src/modules/object-record/record-field/ui/validation-schemas/arrayFieldValueSchema.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const arrayFieldValueSchema = z.union([z.null(), z.array(z.string())]);
