+++ packages/twenty-front/src/modules/object-record/record-field/ui/validation-schemas/fullNameFieldValueSchema.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export const fullNameFieldValueSchema = z.object({
+  firstName: z.string(),
+  lastName: z.string(),
+});
