+++ packages/twenty-front/src/modules/object-record/record-field/ui/validation-schemas/phonesFieldValueSchema.ts
@@ -1,0 +1,18 @@
+import { z } from 'zod';
+
+import { type FieldPhonesValue } from '@/object-record/record-field/ui/types/FieldMetadata';
+
+export const phonesFieldValueSchema = z.object({
+  primaryPhoneNumber: z.string(),
+  primaryPhoneCountryCode: z.string(),
+  primaryPhoneCallingCode: z.string(),
+  additionalPhones: z
+    .array(
+      z.object({
+        number: z.string(),
+        callingCode: z.string(),
+        countryCode: z.string(),
+      }),
+    )
+    .nullable(),
+}) satisfies z.ZodType<FieldPhonesValue>;
