+++ packages/twenty-front/src/modules/object-record/record-field/ui/validation-schemas/richTextFieldValueSchema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+import { type FieldRichTextValue } from '@/object-record/record-field/ui/types/FieldMetadata';
+
+export const richTextFieldValueSchema: z.ZodType<FieldRichTextValue> = z.object({
+  blocknote: z.string().nullable(),
+  markdown: z.string().nullable(),
+});
